pyproject.toml
[project]
name = "meshify-watch"
version = "0.1.0"
description = "Reproducible uv-based wrapper for generating a watch mesh with Hunyuan3D."
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
  "accelerate>=1.9.0",
  "diffusers>=0.35.0",
  "einops>=0.8.1",
  "fastapi>=0.136.0",
  "gradio>=5.42.0",
  "hf-transfer>=0.1.9",
  "huggingface-hub>=0.34.0",
  "hy3dgen",
  "ninja>=1.13.0",
  "numpy>=2.2.6",
  "omegaconf>=2.3.0",
  "onnxruntime>=1.22.0",
  "opencv-python>=4.12.0.88",
  "pillow>=11.3.0",
  "pybind11>=3.0.1",
  "pygltflib>=1.16.5",
  "pymeshlab>=2025.7",
  "python-multipart>=0.0.20",
  "rembg>=2.0.67",
  "safetensors>=0.6.2",
  "torch>=2.8.0",
  "torchvision>=0.23.0",
  "transformers>=4.55.0",
  "trimesh>=4.7.4",
  "tqdm>=4.67.1",
  "uvicorn>=0.35.0",
  "xatlas>=0.0.11",
]

[tool.uv]
package = false

[tool.uv.sources]
hy3dgen = { path = "Hunyuan3D-2", editable = true }
